Fixed error when creating image due to zoom level.
When the Canvas object is created, it has its scaling changed by the code which handles pixel density within flot. When drawing the canvases onto the merged canvas, this scaling would inflate the picture, and cause it to be cropped. mergedContext.restore() fixes this issue.

A second issue occurs when the zoom level changes between creating the plot, and creating the image. The height and width parameters will be different between the canvases, so adjusting the merged canvas height and width fixes this.

The third issue stems from the .find('canvas'), as this would also find theMergedCanvas within the placeholder. Using the .not('.mergedCanvas') fixes this.